+++ 卒業制作作品/卒業制作作品/ObjRailGunAttack.cpp
@@ -33,6 +33,9 @@ void CObjRailGunAttack::Init()
 	//削除距離最大値
 	m_Distance_max = 5;
 
+	//HitBox削除フラグ
+	m_HitBox_Delete = false;
+
 	if (m_RGr == 0 || m_RGr == 180)
 	{
 		//当たり判定用HitBoxを作成
@@ -50,21 +53,13 @@ void CObjRailGunAttack::Init()
 //アクション
 void CObjRailGunAttack::Action()
 {
-	//メニューを開くと停止
+	//メニューを開く、イベント情報表示中は行動停止
 	if (Menu_flg == false)
 	{
-	//位置更新
-	m_RGx += m_RGvx;
-	m_RGy += m_RGvy;
+		//位置更新
+		m_RGx += m_RGvx;
+		m_RGy += m_RGvy;
 	}
-
-	////SE処理
-	//if (Attack_flg == true)
-	//{
-	//	Audio::Start(1); //音楽スタート
-	//	Attack_flg = false; //Attackフラグfalse
-	//}
-
 
 	//主人公位置取得
 	CObjHero* hero = (CObjHero*)Objs::GetObj(OBJ_HERO);
@@ -88,23 +83,19 @@ void CObjRailGunAttack::Action()
 		//主人公から離れるor画面端に行くとオブジェクト削除
 		if (m_RGx < hx - 64 * m_Distance_max)
 		{
-			this->SetStatus(false); //オブジェクト破棄
-			Hits::DeleteHitBox(this); //弾が所有するHitBoxを削除する
+			m_HitBox_Delete = true;
 		}
 		else if (m_RGx > hx + 32 + 64 * m_Distance_max)
 		{
-			this->SetStatus(false); //オブジェクト破棄
-			Hits::DeleteHitBox(this); //弾が所有するHitBoxを削除する
+			m_HitBox_Delete = true;
 		}
 		if (m_RGy < hy - 64 * m_Distance_max)
 		{
-			this->SetStatus(false); //オブジェクト破棄
-			Hits::DeleteHitBox(this); //弾が所有するHitBoxを削除する
+			m_HitBox_Delete = true;
 		}
 		else if (m_RGy > hy + 32 + 64 * m_Distance_max)
 		{
-			this->SetStatus(false); //オブジェクト破棄
-			Hits::DeleteHitBox(this); //弾が所有するHitBoxを削除する
+			m_HitBox_Delete = true;
 		}
 	}
 	
@@ -114,8 +105,7 @@ void CObjRailGunAttack::Action()
 		|| hit_rg->CheckElementHit(ELEMENT_NET_S) == true || hit_rg->CheckElementHit(ELEMENT_NET_V) == true
 		|| hit_rg->CheckElementHit(ELEMENT_BARBED_V) == true)
 	{
-		this->SetStatus(false); //オブジェクト破棄
-		Hits::DeleteHitBox(this); //弾が所有するHitBoxを削除する
+		m_HitBox_Delete = true;
 	}
 	//フィールドエレメントと接触すると削除
 	if (hit_rg->CheckElementHit(ELEMENT_FIELD) == true || hit_rg->CheckElementHit(ELEMENT_FIELD2) == true)
@@ -130,9 +120,17 @@ void CObjRailGunAttack::Action()
 		}
 		else
 		{
-			this->SetStatus(false); //オブジェクト破棄
-			Hits::DeleteHitBox(this); //弾が所有するHitBoxを削除する
+			m_HitBox_Delete = true;
 		}
+	}
+
+	//削除処理
+	if (m_HitBox_Delete == true)
+	{
+		this->SetStatus(false); //オブジェクト破棄
+		Hits::DeleteHitBox(this); //弾が所有するHitBoxを削除する
+
+		m_HitBox_Delete = false; //初期化
 	}
 }
